+++ examples/boilerplate/ar-hello-world/index.html
@@ -2,8 +2,8 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <title>Hello, World! • A-Frame</title>
-    <meta name="description" content="Hello, World! • A-Frame">
+    <title>Hello, AR World! • A-Frame</title>
+    <meta name="description" content="Hello, AR World! • A-Frame">
     <script src="../../../dist/aframe-master.js"></script>
     <script>
       AFRAME.registerComponent('follow-shadow', {
